pipe factory done

+++ BananaYellowProject/src/io/codeforall/kernelfc/Pipe.java
@@ -5,7 +5,7 @@ import org.academiadecodigo.simplegraphics.pictures.Picture;
 
 public class Pipe {
 
-    double y;
+    int y;
     Picture upPipe;
     Picture downPipe;
     // invisible score rectangle
@@ -14,11 +14,11 @@ public class Pipe {
     //gap between the top and bottom pipe, score´s height
     public int gap = 500;
     // tracks the groups X position
-    public double X = 1450;
+    public int X = 1450;
 
     // initiate bottom pipe, top pipe and invisible score rectangle
     // Y parameter to be randomized by the pipe factory
-    public Pipe(double y) {
+    public Pipe(int y) {
         this.y = y;
         upPipe = new Picture(X, y - gap, "resources/toppipe.png");
         downPipe = new Picture(X, y + gap, "resources/bottompipe.png");
@@ -42,7 +42,7 @@ public class Pipe {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
 
@@ -58,7 +58,7 @@ public class Pipe {
         return X;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         X = x;
     }
